Change 'data' from Table to NumPy array (PR #1)
The variable 'data', which contains abundances and is used in the
log-likelihood computation, is converted from an Astropy Table to a
NumPy array. This change means that the function 'lnprob()' can access
the selected abundances without having to explicitly loop through the
names of the chemical elements, which improves performance.